requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.2.1
boto3==1.37.14
botocore==1.37.14
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.8
flake8==7.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
jmespath==1.0.1
mccabe==0.7.0
openai==1.65.2
passlib==1.7.4
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pydantic==2.10.6
pydantic_core==2.27.2
pyflakes==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart
requests==2.32.3
s3transfer==0.11.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.45.3
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0